// catechism/source/types/content.ts
export enum Content {
    PROLOGUE = 'PROLOGUE',
    PART = 'PART',
    SECTION = 'SECTION',
    CHAPTER = 'CHAPTER',
    CHAPTER_SECTION = 'CHAPTER_SECTION',
    ARTICLE = 'ARTICLE',
    ARTICLE_PARAGRAPH = 'ARTICLE_PARAGRAPH',
    SUB_ARTICLE = 'SUB_ARTICLE',
    IN_BRIEF = 'IN_BRIEF',
    PARAGRAPH_GROUP = 'PARAGRAPH_GROUP',

    SPECIAL_1 = 'SPECIAL_1',
    SPECIAL_2 = 'SPECIAL_2',
    SPECIAL_3 = 'SPECIAL_3',

    GENERIC_CONTENT_CONTAINER = 'GENERIC_CONTENT_CONTAINER',

    BLOCK_QUOTE = 'BLOCK_QUOTE',
    PARAGRAPH = 'PARAGRAPH',
    PARAGRAPH_SUB_ITEM_CONTAINER = 'PARAGRAPH_SUB_ITEM_CONTAINER',
    PARAGRAPH_SUB_ITEM = 'PARAGRAPH_SUB_ITEM',
    TEXT_BLOCK = 'TEXT_BLOCK',
    TEXT_HEADING = 'TEXT_HEADING',
    TEXT_WRAPPER = 'TEXT_WRAPPER',
    TEXT = 'TEXT',
}